Rename pawn mock file
